requirements.txt
requests~=2.31.0
openpyxl~=3.1.2
pillow~=10.1.0
nicegui~=1.3.18
pyinstaller~=6.0.0
pywebview~=4.3.3
